Make contract only a passthrough import to export
